+++ modules/gapi/src/backends/common/serialization.hpp
@@ -15,6 +15,7 @@
 
 #include "compiler/gmodel.hpp"
 #include "opencv2/gapi/render/render_types.hpp"
+#include "opencv2/gapi/s11n.hpp" // basic interfaces
 
 #if (defined _WIN32 || defined _WIN64) && defined _MSC_VER
 #pragma warning(disable: 4702)
@@ -30,43 +31,6 @@ struct GSerialized {
     cv::gimpl::DataObjectCounter m_counter;
     cv::gimpl::Protocol m_proto;
 };
-
-////////////////////////////////////////////////////////////////////////////////
-// Stream interfaces, so far temporary
-namespace I {
-    struct GAPI_EXPORTS OStream {
-        virtual ~OStream() = default;
-
-        // Define the native support for basic C++ types at the API level:
-        virtual OStream& operator<< (bool) = 0;
-        virtual OStream& operator<< (char) = 0;
-        virtual OStream& operator<< (unsigned char) = 0;
-        virtual OStream& operator<< (short) = 0;
-        virtual OStream& operator<< (unsigned short) = 0;
-        virtual OStream& operator<< (int) = 0;
-        //virtual OStream& operator<< (std::size_t) = 0;
-        virtual OStream& operator<< (uint32_t) = 0;
-        virtual OStream& operator<< (float) = 0;
-        virtual OStream& operator<< (double) = 0;
-        virtual OStream& operator<< (const std::string&) = 0;
-    };
-
-    struct GAPI_EXPORTS IStream {
-        virtual ~IStream() = default;
-
-        virtual IStream& operator>> (bool &) = 0;
-        virtual IStream& operator>> (char &) = 0;
-        virtual IStream& operator>> (unsigned char &) = 0;
-        virtual IStream& operator>> (short &) = 0;
-        virtual IStream& operator>> (unsigned short &) = 0;
-        virtual IStream& operator>> (int &) = 0;
-        virtual IStream& operator>> (float &) = 0;
-        virtual IStream& operator>> (double &) = 0;
-        //virtual IStream& operator>> (std::size_t &) = 0;
-        virtual IStream& operator >> (uint32_t &) = 0;
-        virtual IStream& operator>> (std::string &) = 0;
-    };
-} // namespace I
 
 ////////////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////
